requirements.txt
tensorflow
numpy
pandas
cobrapy